refactor: color package 이동
